requirements-dev.txt
ansible==2.7.1
pytest==3.10.0
mock==2.0.0
vcrpy-unittest==0.1.7
vcrpy==2.0.1
